pyproject.toml
[project]
name = "asciifier"
version = "0.1.0"
description = "Terminal image-to-character-art CLI with best-fit glyph search and truecolor output"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "pillow>=10.0",
    "numpy>=1.26",
]

[project.scripts]
asciifier = "asciifier.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asciifier"]

[tool.hatch.build.targets.wheel.force-include]
"src/asciifier/fonts/JetBrainsMono-Regular.ttf" = "asciifier/fonts/JetBrainsMono-Regular.ttf"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
